# docs/swagger.yaml
basePath: /
definitions:
  common.PageResp:
    properties:
      content: {}
      total:
        type: integer
    type: object
  common.jsonResult:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  handles.ManualScanReq:
    properties:
      limit:
        type: number
      path:
        type: string
    type: object
...
  handles.SearchReq:
    properties:
      keywords:
        type: string
      page:
        type: integer
      parent:
        type: string
      password:
        type: string
      per_page:
        type: integer
      scope:
        description: 0 for all, 1 for dir, 2 for file
        type: integer
    type: object
  handles.SearchResp:
    properties:
      is_dir:
        type: boolean
      name:
        type: string
      parent:
        type: string
      size:
        type: integer
      type:
        type: integer
    type: object
...
  model.SettingItem:
    properties:
      flag:
        description: 0 = public, 1 = private, 2 = readonly, 3 = deprecated, etc.
        type: integer
      group:
        description: use to group setting in frontend
        type: integer
      help:
        description: help message
        type: string
      index:
        type: integer
      key:
        description: unique key
        type: string
      options:
        description: values for select
        type: string
      type:
        description: string, number, bool, select
        type: string
      value:
        description: value
        type: string
    required:
    - key
    type: object
...
info:
  contact:
    name: OpenList Team
    url: https://github.com/OpenListTeam/
  description: A file list program that supports multiple storage.
  license:
    name: GNU Affero General Public License v3.0
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  termsOfService: https://doc.oplist.org/terms
  title: OpenList
  version: latest
paths:
...
  /api/admin/scan/start:
    post:
      consumes:
      - application/json
      description: Start a manual scan for the specified path with an optional limit
      parameters:
      - description: Manual Scan Data
        in: body
        name: scan
        required: true
        schema:
          $ref: '#/definitions/handles.ManualScanReq'
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Start Manual Scan
      tags:
      - Admin
  /api/admin/scan/stop:
    post:
      description: Stop the ongoing manual scan
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Stop Manual Scan
      tags:
      - Admin
  /api/admin/setting/defaults:
    get:
      consumes:
      - application/json
      description: Get default settings, optionally filtered by group or groups
      parameters:
      - description: Group number to filter settings
        in: query
        name: group
        type: string
      - description: Comma-separated list of group numbers to filter settings
        in: query
        name: groups
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of default setting items
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.SettingItem'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Default Settings
      tags:
      - Admin
  /api/admin/setting/delete:
    post:
      description: Delete a setting by its key
      parameters:
      - description: Key of the setting to delete
        in: query
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Setting
      tags:
      - Admin
  /api/admin/setting/get:
    get:
      consumes:
      - application/json
      description: Get setting by key or keys
      parameters:
      - description: Setting key
        in: query
        name: key
        type: string
      - description: Comma-separated list of setting keys
        in: query
        name: keys
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Single setting item
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  $ref: '#/definitions/model.SettingItem'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Setting
      tags:
      - Admin
  /api/admin/setting/list:
    get:
      consumes:
      - application/json
      description: List settings, optionally filtered by group or groups
      parameters:
      - description: Group number to filter settings
        in: query
        name: group
        type: string
      - description: Comma-separated list of group numbers to filter settings
        in: query
        name: groups
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of setting items
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.SettingItem'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
      summary: List Settings
      tags:
      - Admin
  /api/admin/setting/reset_token:
    post:
      description: Reset the system token and return the new token
      produces:
      - application/json
      responses:
        "200":
          description: Newly generated token
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Reset System Token
      tags:
      - Admin
  /api/admin/setting/save:
    post:
      consumes:
      - application/json
      description: Save multiple settings
      parameters:
      - description: List of setting items to save
        in: body
        name: settings
        required: true
        schema:
          items:
            $ref: '#/definitions/model.SettingItem'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Save Settings
      tags:
      - Admin
...
  /api/fs/search:
    post:
      consumes:
      - application/json
      description: Search files and folders within the user's accessible scope
      parameters:
      - description: Search request
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/handles.SearchReq'
      produces:
      - application/json
      responses:
        "200":
          description: Search results with pagination
          schema:
            allOf:
            - $ref: '#/definitions/common.PageResp'
            - properties:
                content:
                  items:
                    $ref: '#/definitions/handles.SearchResp'
                  type: array
                total:
                  type: integer
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
      security:
      - Authorization: []
      summary: Search Files and Folders
      tags:
      - Filesystem
...
  /api/public/settings:
    get:
      consumes:
      - application/json
      description: Get all public settings as a key-value map
      produces:
      - application/json
      responses:
        "200":
          description: Map of public settings
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Public Settings
      tags:
      - Public
...
securityDefinitions:
  Authorization:
    description: Token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"